--- Ray.py
@@ -1,139 +1,0 @@
-import math
-
-class Ray():
-    rays = []
-    def __init__(self, start_coords, end_coords) -> None:
-        self.x0, self.y0 = start_coords
-        self.x1, self.y1 = end_coords
-        Ray.rays.append(self)
-
-    def get_start_coords(self):
-        return (self.x0, self.y0)
-
-    def get_end_coords(self):
-        return (self.x1, self.y1)
-
-    def set_start_coords(self, new_start):
-        self.x0, self.y0 = new_start
-
-    def set_end_coords(self, new_end):
-        (self.x1, self.y1) = new_end
-
-    def get_reflected_ray(self, ray_obj, angle):
-        return Ray(self.get_start_coords(), self.get_end_coords())
-
-    def get_delta_x(self):
-        return abs(self.x1 - self.x0)
-
-    def get_delta_y(self):
-        return abs(self.y1 - self.y0)
-
-    def get_slope_intercept_form(self):
-        x, y = self.get_start_coords()
-        m = self.get_slope()
-        b = y - m*x
-        return (m, b)
-
-    def get_length(self):
-        return (self.get_delta_x(), self.get_delta_y())
-
-    def get_center_coords(self):
-        return ((self.x0 + self.x1)/2, (self.y0 + self.y1)/2)
-
-    def get_angle(self):
-        return math.atan2(self.get_delta_x(), self.get_delta_y())
-
-    def get_slope(self):
-        return (self.y1-self.y0) / (self.x1-self.x0)
-
-
-    def rotate(self,pivot=None, angle=None):
-        if angle == None:
-            angle = math.pi/2
-        new_x0, new_y0 = self.get_start_coords()
-        new_x1, new_y1 = self.get_end_coords()
-        new_x_mid, new_y_mid = self.get_center_coords()
-        if pivot == (new_x0, new_y0) or pivot == None:
-            trans_x0 = new_x0 - new_x0
-            trans_x1 = new_x1 - new_x0
-            trans_y0 = new_y0 - new_y0
-            trans_y1 = new_y1 - new_y0
-        elif pivot == (new_x1, new_y1):
-            trans_x0 = new_x0 - new_x1
-            trans_x1 = new_x1 - new_x1
-            trans_y0 = new_y0 - new_y1
-            trans_y1 = new_y1 - new_y1
-        elif pivot == (new_x_mid, new_y_mid):
-            trans_x0 = new_x0 - new_x_mid
-            trans_x1 = new_x1 - new_x_mid
-            trans_y0 = new_y0 - new_y_mid
-            trans_y1 = new_y1 - new_y_mid
-
-        self.x0 = trans_x0*math.cos(angle) - trans_y0*math.sin(angle)
-        self.x1 = trans_x1*math.cos(angle) - trans_y1*math.sin(angle)
-        self.y0 = trans_x0*math.sin(angle) + trans_y0*math.cos(angle)
-        self.y1 = trans_x1*math.sin(angle) + trans_y1*math.cos(angle)
-
-        if pivot == (new_x0, new_y0) or pivot == None:
-            self.x0 += new_x0
-            self.x1 += new_x0
-            self.y0 += new_y0
-            self.y1 += new_y0
-        elif pivot == (new_x1, new_y1):
-            self.x0 += new_x1
-            self.x1 += new_x1
-            self.y0 += new_y1
-            self.y1 += new_y1
-        elif pivot == (new_x_mid, new_y_mid):
-            self.x0 += new_x_mid
-            self.x1 += new_x_mid
-            self.y0 += new_y_mid
-            self.y1 += new_y_mid
-
-    def copy(self):
-        return Ray(self.get_start_coords(), self.get_end_coords())
-
-    def move(self, destination_coord):
-        movement_coords = (destination_coord[0] - self.get_start_coords()[0], destination_coord[1] - self.get_start_coords()[1])
-        new_start_coords = (self.get_start_coords()[0] + movement_coords[0], self.get_start_coords()[1] + movement_coords[1])
-        new_end_coords = (self.get_end_coords()[0] + movement_coords[0], self.get_end_coords()[1] + movement_coords[1])
-        self.set_start_coords(new_start_coords)
-        self.set_end_coords(new_end_coords)
-
-    def move_vertical(self, y_amount):
-        self.y0 += y_amount
-        self.y1 += y_amount
-    def move_horizontal(self, x_amount):
-        self.x0 += x_amount
-        self.x1 += x_amount
-
-    def extend(self, room_size):
-        x0, y0 = self.get_start_coords()
-        x1, y1 = self.get_end_coords()
-        # going left
-        if x0 > x1:
-            x_towidth = 0
-            m_towidth, b_towidth = self.get_slope_intercept_form()
-            y_towidth = m_towidth*x_towidth + b_towidth
-        # going right
-        elif x0 < x1:
-            x_towidth = room_size[0]
-            m_towidth, b_towidth = self.get_slope_intercept_form()
-            y_towidth = m_towidth*x_towidth + b_towidth
-        #going down
-        if y0 < y1:
-            y_toheight = room_size[1]
-            m_toheight, b_toheight = self.get_slope_intercept_form()
-            x_toheight = (y_toheight - b_toheight)/m_toheight
-        #going up
-        elif y0 > y1:
-            y_toheight = 0
-            m_toheight, b_toheight = self.get_slope_intercept_form()
-            x_toheight = (y_toheight - b_toheight)/m_toheight
-
-        len_toheight = ((y_toheight-y0)**2 + (x_toheight-x0)**2)**0.5
-        len_towidth = ((y_towidth-y0)**2 + (x_towidth-x0)**2)**0.5
-        if len_toheight < len_towidth:
-            self.set_end_coords((x_toheight, y_toheight))
-        else:
-            self.set_end_coords((x_towidth, y_towidth))
